todo 1-8
TODO 1:-
added background image to the app background in container using decoration: box decoration property.

TODO 2:-
created a new class story to have the storytitle and choices.

TODO 3:-
added the properties to this class.

TODO 4:-
created a cosntructor for this.

TODO 5:-
created a new class story_brain.
and imported story.dart

TODO 7:-
included the list in the class.

TODO 8:-
created a getter method getStory() to return the story title.